calc: do not request/show default view area tiles before...
we get the cell cursor position from core which determines the final
view area. If we do on document load the default view starting at A1 is
displayed for a short time and after getting the cursor position, new
tiles are requested for the new view area and are displayed. This is
also perceived as a flicker on load.

Also request the cell cursor position early during the addition of
CalcTileLayer to map.

Signed-off-by: Dennis Francis <dennis.francis@collabora.com>
Change-Id: I0996b75ed120a59a242cb50991d90afd228d3582