namespace mount can fail due to adding MS_NOATIME
in deployment on remounting the mount pointing to /opt/cool/systemplate
with an additional MS_NOATIME results in EPERM.  Where that dir is on a
(toplevel) [rel]atime mount.

man 2 mount states 'An attempt was made to modify (MS_REMOUNT) the
MS_RDONLY, MS_NOSUID, or MS_NOEXEC flag, or one of the "atime" flags
(MS_NOATIME, MS_NODIRATIME, MS_RELATIME) of an existing mount, but the
mount is locked'. Presumably we can add flags that drop privs, but not
those that could circumvent original mount policy.

Thanks to Kay Sievers for help & advice to bootstrap these namespace
mounts.

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: I4c0c6a6e4a0e7fae04255e247b18cd5a86c3f327