Render current slide, not just when we tranition
This change adds rendering of the current slide when the image
becomes available. So rendering of the sldie is also performed
when we don't transition from one slide to the other.

To make this possible part of Transition2d code was factored
out to a new RenderContext class and the SlideRenderer is
responsible to set up the rendering...

Also change the transitions to use a class transitionParameter
instead of a bunch of mostly the same parameters used by the
transition sub-classes.

Signed-off-by: Tomaž Vajngerl <tomaz.vajngerl@collabora.co.uk>
Change-Id: If670baf9e6dbe273e4a68d1c9cc0c5bad7bb604d