render video on top of the slide, if there is any declared
Adds video playback as a texture in WebGL. if there is some video
declared in the presentationInfo JSON from core and is accesible
through the public url. For such videos we create a video element
which can be then added as a texture to the WebGL and rendered on
top of the slide.

Signed-off-by: Tomaž Vajngerl <tomaz.vajngerl@collabora.co.uk>
Change-Id: Ie0b003e98c566c560280e8bf2dfbcc408d7d93f7